tests: Paralellize gitops tests
